remove _layout.tsx file and rename index.tsx to RequiredInput.tsx
